refactoring direct  logic
