+++ backend/internal/entities/tables.go
@@ -1,7 +1,6 @@
 package entities
 
 import (
-	"backend/modules/user/models"
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 	"time"
@@ -44,39 +43,5 @@ func (attempt *LoginAttempt) BeforeCreate(*gorm.DB) error {
 	if attempt.ID == uuid.Nil {
 		attempt.ID = uuid.New()
 	}
-	return nil
-}
-
-type DirectMessage struct {
-	ID             uuid.UUID `gorm:"type:uuid;primaryKey" json:"id"`
-	ConversationID uuid.UUID `gorm:"not null"`
-	SenderID       uuid.UUID `gorm:"type:uuid;not null"`
-	Text           string    `gorm:"type:text;not null"`
-	Read           bool      `gorm:"default:false"`
-	CreatedAt      time.Time
-
-	Sender       models.User   `gorm:"foreignKey:SenderID"`
-	Conversation Conversations `gorm:"foreignKey:ConversationID"`
-}
-
-func (dm *DirectMessage) BeforeCreate(*gorm.DB) error {
-	dm.ID = uuid.New()
-	return nil
-}
-
-type Conversations struct {
-	ID        uuid.UUID `gorm:"type:uuid;primaryKey" json:"id"`
-	User1ID   uuid.UUID `gorm:"type:uuid;not null"`
-	User2ID   uuid.UUID `gorm:"type:uuid;not null"`
-	CreatedAt time.Time
-
-	User1 models.User `gorm:"foreignKey:User1ID"`
-	User2 models.User `gorm:"foreignKey:User2ID"`
-
-	DirectMessage []DirectMessage `gorm:"foreignKey:ConversationID"`
-}
-
-func (c *Conversations) BeforeCreate(*gorm.DB) error {
-	c.ID = uuid.New()
 	return nil
 }
